adding toast, create service folder
